fix: calculation of avg price for offers (#625)
* fix: calculation of avg price for offers

* chore: remove total-capacity filter from offer

* chore: adjust tests

+++ src/services/storage/models/offer.model.ts
@@ -96,7 +96,7 @@ export function getBillingPriceAvgQuery (
         SUM(
           (cast(price as REAL) / ${WEI}) * coalesce("rates".${sequelize.escape(currency)}, 0) * 1024 / period * (3600 * 24 * 30)
         ) / COUNT("storage_billing_plan"."id")
-        as INTEGER
+        as REAL
       )
     FROM
       "storage_billing_plan"
